Move spinner to view
