+++ client/analysis.ts
@@ -5,7 +5,7 @@ import { AnalysisController } from './analysisCtrl';
 import { gameInfo } from './gameInfo';
 import { selectVariant, VARIANTS } from './variants';
 import { renderTimeago } from './datetime';
-import { spinner } from './spinner';
+import { spinner } from './view';
 import { PyChessModel } from "./types";
 
 function runGround(vnode: VNode, model: PyChessModel) {
